feat: LCIN-950 Node removal/addition in syngraph (#42)
The functions for add and remove reaction nodes from SynGraph instances have coherent APIs and in both cases it is allowed to work only with reaction nodes (not molecules).

The SynGraph method to remove single node is improved and modified to work based on node IDs instead of full ChemicalEquation/Molecule instances.

Few additional improvements are implemented.