LCIN.1211: Route metrics refactoring (#94)
The route metrics module is refactored:

each metric class now is required to have a method to extract the "components" (molecules or chemical equations) for which data are needed in order to compute the metric. This method output a RouteComponents object which stores the type of component (molecules/chemicalequations), a map between the uid of the components and their structural representation and the type of structural representetion (smiles, smarts, etc..)
the method to actually compute the metric is now expecting a dictionary whose keys are the uid of the component and whose values the external data
some utilities functions/structures are moved to a new submodule to keep the metric module more readable
The CI/CD pipeline is updated to avoid using a deprecated os.

Flake8 linting is applied to some modules within the rem submodule.