refactor: CAST-500 Constructors refactoring (#103)
The Constructors of the cheminformatics data models are refactored to make the code more readable and maintainable.
Improvements to the chemical_hashes module are implemented.

The fragmented_rdmol attribute of the Disconnection model is now excluded from the main workflow to create the object.
The Template object is no longer created when a ChemicalEquation object is instantiated. The legacy code is still available in a dedicated module.